refacto DB (wip)
