+++ enjoliver-api/tests/unit/test_machine_discovery_repo.py
@@ -1,9 +1,12 @@
 import copy
 import unittest
 
-from enjoliver import smartdb, model
-from enjoliver.model import MachineInterface, Machine, MachineDisk, Chassis, ChassisPort
-from enjoliver.repositories.machine_discovery_repo import DiscoveryRepository
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
+from enjoliver.db import session_commit
+from enjoliver.model import MachineInterface, Machine, MachineDisk, Chassis, ChassisPort, Base
+from enjoliver.repositories.machine_discovery import MachineDiscoveryRepository
 
 from tests.fixtures import posts
 
@@ -11,26 +14,26 @@ from tests.fixtures import posts
 class TestMachineStateRepo(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        db_uri = 'sqlite:///:memory:'
-
-        cls.smart = smartdb.SmartDatabaseClient(db_uri)
+        db_uri = 'postgresql+psycopg2://localhost/enjoliver_testing'
+        cls.engine = create_engine(db_uri)
+        cls.sess_maker = sessionmaker(bind=cls.engine)
 
     def setUp(self):
-        model.BASE.metadata.drop_all(self.smart.get_engine_connection())
-        model.BASE.metadata.create_all(self.smart.get_engine_connection())
+        Base.metadata.drop_all(bind=self.engine)
+        Base.metadata.create_all(bind=self.engine)
 
     def test_bad_content(self):
-        mdr = DiscoveryRepository(self.smart)
+        mdr = MachineDiscoveryRepository(self.sess_maker)
         with self.assertRaises(TypeError):
             mdr.upsert(dict())
         with self.assertRaises(TypeError):
             mdr.upsert({"lldp": ""})
 
     def test_no_machine(self):
-        mdr = DiscoveryRepository(self.smart)
+        mdr = MachineDiscoveryRepository(self.sess_maker)
         mdr.upsert(posts.M01)
 
-        with self.smart.new_session() as session:
+        with session_commit(sess_maker=self.sess_maker) as session:
             self.assertEqual(1, session.query(Machine).count())
             self.assertEqual(1, session.query(MachineInterface).count())
             self.assertEqual(1, session.query(MachineDisk).count())
@@ -38,12 +41,12 @@ class TestMachineStateRepo(unittest.TestCase):
             self.assertEqual(1, session.query(ChassisPort).count())
 
     def test_no_machine_readd_same(self):
-        mdr = DiscoveryRepository(self.smart)
+        mdr = MachineDiscoveryRepository(self.sess_maker)
 
         for i in range(3):
             mdr.upsert(posts.M01)
 
-            with self.smart.new_session() as session:
+            with session_commit(sess_maker=self.sess_maker) as session:
                 self.assertEqual(1, session.query(Machine).count())
                 self.assertEqual(1, session.query(MachineInterface).count())
                 self.assertEqual(1, session.query(MachineDisk).count())
@@ -51,10 +54,10 @@ class TestMachineStateRepo(unittest.TestCase):
                 self.assertEqual(1, session.query(ChassisPort).count())
 
     def test_no_machine_readd_disk_diff(self):
-        mdr = DiscoveryRepository(self.smart)
+        mdr = MachineDiscoveryRepository(self.sess_maker)
         mdr.upsert(posts.M01)
 
-        with self.smart.new_session() as session:
+        with session_commit(sess_maker=self.sess_maker) as session:
             self.assertEqual(1, session.query(Machine).count())
             self.assertEqual(1, session.query(MachineInterface).count())
             self.assertEqual(1, session.query(MachineDisk).count())
@@ -65,7 +68,7 @@ class TestMachineStateRepo(unittest.TestCase):
         with_new_disk["disks"].append({'size-bytes': 21474836481, 'path': '/dev/sdb'})
         mdr.upsert(with_new_disk)
 
-        with self.smart.new_session() as session:
+        with session_commit(sess_maker=self.sess_maker) as session:
             self.assertEqual(1, session.query(Machine).count())
             self.assertEqual(1, session.query(MachineInterface).count())
             self.assertEqual(2, session.query(MachineDisk).count())
@@ -73,10 +76,10 @@ class TestMachineStateRepo(unittest.TestCase):
             self.assertEqual(1, session.query(ChassisPort).count())
 
     def test_no_machine_remove_disks(self):
-        mdr = DiscoveryRepository(self.smart)
+        mdr = MachineDiscoveryRepository(self.sess_maker)
         mdr.upsert(posts.M01)
 
-        with self.smart.new_session() as session:
+        with session_commit(sess_maker=self.sess_maker) as session:
             self.assertEqual(1, session.query(Machine).count())
             self.assertEqual(1, session.query(MachineInterface).count())
             self.assertEqual(1, session.query(MachineDisk).count())
@@ -87,7 +90,7 @@ class TestMachineStateRepo(unittest.TestCase):
         without_disks["disks"] = None
         mdr.upsert(without_disks)
 
-        with self.smart.new_session() as session:
+        with session_commit(sess_maker=self.sess_maker) as session:
             self.assertEqual(1, session.query(Machine).count())
             self.assertEqual(1, session.query(MachineInterface).count())
             self.assertEqual(0, session.query(MachineDisk).count())
@@ -95,7 +98,7 @@ class TestMachineStateRepo(unittest.TestCase):
             self.assertEqual(1, session.query(ChassisPort).count())
 
     def test_fetch_one_machine(self):
-        mdr = DiscoveryRepository(self.smart)
+        mdr = MachineDiscoveryRepository(self.sess_maker)
         mdr.upsert(posts.M01)
 
         disco = mdr.fetch_all_discovery()
